+++ ruma-federation-api/src/membership/create_join_event.rs
@@ -1,6 +1,7 @@
 //! Endpoint to send join events to remote homeservers.
 
 pub mod v1;
+pub mod v2;
 
 use ruma_common::Raw;
 use ruma_events::pdu::Pdu;