+++ ruma-federation-api/src/membership/create_join_event/v2.rs
@@ -2,6 +2,7 @@
 
 use ruma_api::ruma_api;
 use ruma_events::pdu::PduStub;
+use ruma_common::Raw;
 use ruma_identifiers::{EventId, RoomId};
 
 use super::RoomState;
@@ -26,11 +27,12 @@ ruma_api! {
 
         /// PDU type without event and room IDs.
         #[ruma_api(body)]
-        pub pdu_stub: PduStub,
+        pub pdu_stub: Raw<PduStub>,
     }
 
     response: {
         /// Full state of the room.
+        #[ruma_api(body)]
         pub room_state: RoomState,
     }
 }